selinux: move the package under sandbox/selinux
Move the selinux package to be under sandbox, where other sandbox related pieces
live.
